Created CPL_Imager.py, the entry point to the program and the file that hosues the main class. Basically just initialises the SDK and opens the right cameras and creates the widgets etc. Have made a 3 classes: a two camera mode, a one camera mode and a no camera (mock) mode.
Have also made a GUI class that houses all the TKinter button logic and packing arrangements etc. For some reason seems to work best when both the GUI and the canvas are packed into root.

Next commit will be refactoring the code - placing the mock camera and the LiveViewCamera from live view into a file called cameras.py and getting rid of the colour code and script stuff, and throwing the LiveView canvas into the GUI section where it more naturally belongs. This should leave a nice 4 file structure that's easily understandable.